requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing_extensions==4.2.0
whitenoise==6.1.0
